GetMovieById, IncludableQuery

+++ BestMovies.Net/BestMovies.Api/Models/Rating.cs
@@ -3,7 +3,7 @@
 public class Rating
 {
     public int MovieId { get; set; }
-    public double Value { get; set; }
+    public float Value { get; set; }
     public int Votes { get; set; }
 
     public virtual Movie Movie { get; set; } = null!;
